aws/src/auth/tests/requirements.txt
boto3
moto
bcrypt
PyJWT
